Rework keyboard interface
This commit greatly reworks the keyboard interface, making the code
cleaner and more easily maintainable in the future. A lot of useful
functions and initialization is now part of the AbstractKeyboard, found
in `abc.py`, instead of being duplicated across almost all modules.

The dummy keyboard module is no longer necessary and has been removed.
What was previously known as "multi-keyboard mode" is now simply the
default (and only supported) modus operandi.

The section headers in the configuration file have been changed to make
more clear what they are about. The default `doorpi.ini` in the `data`
subdirectory has been updated accordingly; please refer to it for info
about the specific changes.