doorpi: Rewrite main loop for synchronous ticks
OnTimeTick events are now fired synchronously from the main thread
(instead of being fired asynchronously, as other OnTime* events are).
Additionally, OnTimeRapidTick events are fired synchronously every 50ms.
Instead of being hard-coded, the regular self checks for the SIP phone
module must now be registered as actions for OnTimeRapidTick or
OnTimeTick, depending on the needs of the specific module.

Being synchronous in the main thread, tick events should no longer be
used for long running operations. Instead, use one of the other OnTime*
events, which are still asynchronous.

If a tick takes longer than allowed by the tick rate (50ms), a warning
message will be emitted, along with information on which actions are
registered for both tick events. If that happens and multiple ticks have
to be skipped, it is possible for some OnTime* events to be lost.