doorpi: Remove restart() method
Restarting without terminating the process could lead to some weird and
hard-to-debug edge cases, which is why it is now being removed. Instead,
shut down the program normally (e.g. by sending it SIGTERM) and then
launch another instance of it. This works best if DoorPi is running
under a daemontools-like supervisor like systemd.