main: Allow setting debug/trace logging only for specific modules
This change allows for greater flexibility while debugging the
application. Setting only a defined subset of modules to DEBUG or TRACE
logging greatly reduces the amount of unrelated log messages. This
allows to focus better on the task at hand.

With this change, both `--trace` and `--debug` accept an optional
argument, which names a specific logger to operate on (e.g.
`doorpi.main`). If no argument is specified, the change is applied
globally; this means that `doorpi_cli --trace` will still work as
before.

`--trace` and `--debug` may be specified multiple times, in which case
all mentioned modules will respect the specific log level. Regardless of
the order on the command line, a `--trace` will override a `--debug` for
the same module.