Improve BasePath finding
On Linux, the XDG base directory specification is honored. On Windows,
the more conventional `%APPDATA%\DoorPi` is used. Other systems now use
to `~/.doorpi`. In any case, if the previously default `~/doorpi.ini` is
found, the old default of `~` will still be used. This is to avoid
breaking existing setups.