Change how the DoorPi instance is found from client code
Instead of a custom `Singleton` meta-class that keeps a hidden reference
to the instance alive continuously, create an `INSTANCE` constant in the
top-level `doorpi` module. Upon creating the instance, it will be
populated with a reference to it, and when the instance is being
destroyed the reference will be removed again too.