web: Refactor web server using aiohttp
The previous implementation using an OS thread per client connection
could not provide information about PJSIP calls.  The PJSIP library
expects all threads that want to use it to be registered first.
Registering permanently allocates an amount of memory, which makes it
unsuitable for a web server that uses ephemeral threads.  Switching to
the aiohttp framework does away with the issue by using only a single OS
thread for all requests, which can be registered with PJSIP without
causing a memory leak.