keyboard: Ensure unique enum member names
The entry points logic doesn't care about names being defined multiple
times, but for enums this is an error.  While duplicates shouldn't
happen under normal circumstances, there are some cases where they can
legitimately occur.  Most notably, when the `doorpi` package was
installed using `pip install`, and the test suite is run inside a
different source tree.  In this case, both locations define the same
entry point names (with the same value), which leads to clashes.

Using a dict instead of an iterator of 2-tuples for the `Enum` type
constructor circumvents the issue, because the dict will only retain the
last entry for each name.