actions: Add `waitevent`
The waitevent can be used to conditionally abort or continue executing
an event, based on whether or not another event occured within a given
time period.

The action will halt event execution for the given time period and wait
for the named event to occur.  Event execution is resumed (or aborted)
once the timeout expired, or immediately after the named other event
occured.